Moved utility class
- We have move an old utility class and given it a better name.
